requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
sib-api-v3-sdk
streamlit
requests

